added some fields to product list items

+++ src/screens/ArtisanHub/actions.js
@@ -123,8 +123,11 @@ export function fetchProducts(artisanID) {
 export function deleteArtisan(artisans, artisan) {
   return (dispatch) => {
     return new Promise(async (resolve, reject) => {
+      const artisanToDelete = artisans.find((item) => item.uid === artisan);
+      if(artisanToDelete.profilePictureURL) {
+         await firebase.storage().ref(`artisanFiles/${artisan}/images/profilePicture`).delete();
+      }
       await firebase.database().ref(`artisans/${artisan}`).remove()
-      await firebase.storage().ref(`artisanFiles/${artisan}/images/profilePicture`).delete()
       
       resolve()
       dispatch({type: 'DELETE_ARTISAN', artisan: artisan})
